+++ backend/app/Models/User.php
@@ -47,6 +47,16 @@ class User extends Authenticatable
         return $this->hasMany(Project::class);
     }
 
+    public function followers()
+    {
+        return $this->hasMany(Follow::class, 'following_id');
+    }
+
+    public function following()
+    {
+        return $this->hasMany(Follow::class, 'follower_id');
+    }
+
     /**
      * Get the attributes that should be cast.
      *